test(cli): move auto-update config precedence coverage
