+++ packages/cli/test/config.test.ts
@@ -25,6 +25,7 @@ import {
   sharedHomeInitGate,
   repoDir,
   resolveNetworkConfigName,
+  resolveAutoUpdateConfig,
   resolveAutoUpdateSource,
   resolveApprovalPolicy,
   resolveChainConfig,
@@ -665,6 +666,55 @@ describe('resolveAutoUpdateSource', () => {
       { autoUpdate: { enabled: false } },
       { autoUpdate: { enabled: false } as any },
     )).toBeUndefined();
+  });
+});
+
+describe('resolveAutoUpdateConfig', () => {
+  it('inherits network-level tag-signature verification for git auto-update polling', () => {
+    const resolved = resolveAutoUpdateConfig(
+      {
+        autoUpdate: {
+          enabled: true,
+          source: 'git',
+        },
+      },
+      {
+        autoUpdate: {
+          enabled: true,
+          repo: 'owner/repo',
+          branch: 'main',
+          checkIntervalMinutes: 30,
+          verifyTagSignature: true,
+          source: 'git',
+        },
+      },
+    );
+
+    expect(resolved?.verifyTagSignature).toBe(true);
+  });
+
+  it('preserves a local false tag-signature override over network defaults', () => {
+    const resolved = resolveAutoUpdateConfig(
+      {
+        autoUpdate: {
+          enabled: true,
+          source: 'git',
+          verifyTagSignature: false,
+        },
+      },
+      {
+        autoUpdate: {
+          enabled: true,
+          repo: 'owner/repo',
+          branch: 'main',
+          checkIntervalMinutes: 30,
+          verifyTagSignature: true,
+          source: 'git',
+        },
+      },
+    );
+
+    expect(resolved?.verifyTagSignature).toBe(false);
   });
 });
